requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
click==8.1.8
et_xmlfile==2.0.0
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
